backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.5
jinja2==3.1.4
python-multipart==0.0.9
pydantic-settings==2.4.0
httpx==0.27.2
redis==5.0.7
rq==1.16.2
tenacity==8.5.0
sqlalchemy>=2.0
psycopg[binary]==3.2.1
alembic==1.13.2
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.2.0
redis>=5.0
pydantic>=2.7
pydantic-settings>=2.2
python-jose[cryptography]>=3.3